Ficha reserva y ficha espacio listas

+++ cc4401Inventory/reservationsApp/views.py
@@ -1,6 +1,8 @@
 from django.shortcuts import render, redirect
 from .models import Reservation
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+
 
 def reservation_data(request, reservation_id):
     try:
@@ -27,13 +29,18 @@ def delete(request):
 
         return redirect('user_data', user_id=request.user.id)
 
+@login_required
 def cancel(request):
     if request.method == 'POST':
         reservation_id = request.POST.get('reservation')
         reservation = Reservation.objects.get(id=reservation_id)
         try:
-            if reservation.state == 'P':
-                reservation.delete()
+            user = request.user
+            if reservation.user == user:
+                if reservation.state == 'P':
+                    reservation.delete()
+            else:
+                messages.warning(request, 'Ud no puede cancelar esta reserva')
         except:
             messages.warning(request, 'Ha ocurrido un error y la reserva no se ha cancelado')
